requirements.txt
Flask 
python-telegram-bot